Align with R2DBC 0.9 API changes
Adopt to R2DBC Parameter type, deprecate our own one in favor of the R2DBC type.
Add support for extensible transaction definitions and support to consume Readable and result segments.
Return Long instead of Integer in DatabaseClient update in preparation for R2DBC 1.0 changes.

+++ spring-r2dbc/src/main/kotlin/org/springframework/r2dbc/core/UpdatedRowsFetchSpecExtensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2020 the original author or authors.
+ * Copyright 2002-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,5 +23,5 @@ import kotlinx.coroutines.reactive.awaitSingle
  *
  * @author Fred Montariol
  */
-suspend fun UpdatedRowsFetchSpec.awaitRowsUpdated(): Int =
+suspend fun UpdatedRowsFetchSpec.awaitRowsUpdated(): Long =
 		rowsUpdated().awaitSingle()
